ARROW-7091: [C++] Move DataType factory decls to type_fwd.h
